+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Get-ForgeModuleInfo` function for retrieving releases and PowerShell metadata about Puppetized PowerShell modules on the forge ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
 - `Get-PowerShellDscModule` function for discovering Puppetizable PowerShell modules in a PowerShell repository ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
 
+### Changed
+
+- `Publish-PuppetModule` function to include the new `ExportFolder`, `PackagedModulePath`, and `Force` parameters, as well as refactor the function to use `Export-PuppetModule` and clean up the experience simultaneously building and publishing a Puppet module. Finally, the `ForgeToken` parameter now defaults to the `FORGE_TOKEN` environment variable if it is not passed directly ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
+
 ### Fixed
 
 - Ensure that PowerShell module's missing the `ProjectURI` metadata key are still Puppetizable; they are valid modules on the PowerShell Gallery, so default the 'file an issue' link to the module's gallery link if the `ProjectUri` is missing ([#186](https://github.com/puppetlabs/ruby-pwsh/issues/186))